Created separate components and moved files around
