+++ client/components/SelfCheck.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const SelfCheck = () => (
+    <div className="textComponent">
+       <img className="topicIcons" src="../img/self-check.JPG"></img>
+       <div className="wordish">
+       <h3>Self Check-in</h3>
+       <p>Check yourself in with the keypad.</p>
+       </div>
+     </div>
+)
+
+export default SelfCheck;
