+++ client/components/SuperHost.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const SuperHost = () => (
+    <div className="textComponent">
+       <img className="topicIcons" src="../img/host.JPG"></img>
+       <div className="wordish">
+       <h3>Hilary is a Superhost</h3>
+       <p>Superhosts are experienced, highly rated hosts who are commited to providing great stays for guests.</p>
+       </div>
+     </div>
+)
+
+export default SuperHost;
