+++ client/components/hosted.jsx
