+++ package.json
@@ -22,6 +22,7 @@
     "react-dom": "^17.0.1",
     "save-dev": "0.0.1-security",
     "style-loader": "^2.0.0",
+    "webpack": "^5.17.0",
     "webpack-cli": "^4.4.0"
   },
   "devDependencies": {
@@ -30,7 +31,6 @@
     "babel-loader": "^8.2.2",
     "css-loader": "^5.0.1",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
-    "express": "^4.17.1",
-    "webpack": "^5.16.0"
+    "express": "^4.17.1"
   }
 }
